Initialize GitHub Pages dummy branch
This commit serves to initialize a new GitHub Pages dummy branch for the
GitHub Pages built-in workflow. By configuring this workflow to be
triggered by changes to this branch, we can effectively disable the
workflow without also disabling GitHub Pages on this repository.

A sole CNAME file is present so that GitHub Pages uses the indicated
custom domain.