Make Jekyll theme local to project
Converts the current Jekyll theme (Minima) to a local gem by copying all
of the Minima v2.5.1 source files into the new _theme directory. With
this theme now local to the project, it will be easier to further
customize the website.

Over time, I would like to customize this website with my own home-grown
theme so that it is unique. I plan on doing so by transitioning away
from Minima iteratively.