Remove X-UA-Compatible meta tag from default head
Removes the "X-UA-Compatible" meta tag from the site's default head
template. It appears after some superficial investigation (see, for
example, answers on StackOverflow such as
https://stackoverflow.com/a/6771584 and
https://stackoverflow.com/a/26348511 advising against keeping this tag
around) that this tag is not something that is needed anymore in
the modern internet, hence I am removing it in favor of not cluttering
the website.