Configure body to not be smaller than its elements
Configures the site body to use the CSS grid layout, which has the side
effect of making it so that the body doesn't shrink smaller than its
elements when the viewport is collapsed horizontally.

This change may be a bit of a hack. I am not sure, as I don't understand
some of the nuances of the CSS grid layout.