Replace raw HTML in program usage examples
Replaces the raw HTML snippets for the two program usage examples in the
2022-04-11 blog post with Markdown code blocks with an associated id.
The Javascript programs associated with these code snippets were also
necessarily modified to use CSS path selectors.

This change significantly simplifies the Markdown source without making
any changes to the rendered site. The only reason that the previous
method was used was because I was not previously fully aware that it was
possible.