Enable displaying language on code blocks
Adds styling that enables the displaying of a language descriptor on
code blocks by default. Also adds the ability to hide this descriptor on
an individual basis via the use of a CSS class (namely,
'.language-descriptor-hidden'), and hides this descriptor on an
assortment of code blocks for which the descriptor was deemed
counterproductive.