Switch implementation of main margins
Switches the implementation of how margins between the header, main, and
footer elements of the site are created. Apparently, margins of children
of elements with a grid layout do not collapse, so they have to be
manually removed to acheive a similar effect.

See https://stackoverflow.com/a/37837971 for more information on the
lack of collapsing margins in grid layouts.