Add empty syntax highlighting template stylesheet
Adds and enables a stylesheet for syntax highlighting code blocks (and
spans). This is effectively a blank template, although it contains a
list of all standard Pygments token types. I intend on eventually
customizing syntax highlighting via this stylesheet as necessary.

The list of all token types was originally derived from
cross-referencing the following two pieces of documentation:
https://pygments.org/docs/tokens and
https://github.com/pygments/pygments/blob/2.12.0/pygments/token.py.
For some reason, neither of these two references contains a
comprehensive list of all tokens, but after some investigation it
appears that the combination of both is comprehensive.