Improve implementation of custom HTML converter
Switches the implementation of the customizations to Kramdowns HTML
converter from direct monkey patching to using Module#prepend. This new
implementation is apparently considered a better practice than direct
monkey patching, and it is slightly cleaner as well.

See https://stackoverflow.com/a/4471202 for an overview on the various
alternatives to direct monkey patching in Ruby.