Add jekyll-figure dependency
Adds the jekyll-figure dependency, which allows for rendering HTML
figure elements using Liquid expressions.

Note that this plugin works best when used in conjunction with the HTML
converter customization added in commit
e8373aa77d823c00d91cac3e58a84a3eec677d92. This is because Kramdown by
default renders standalone elements wrapped in paragaphs, but this leads
to the creation of superfluous paragraphs inside of figure elements by
this plugin. Using these two plugins in conjunction avoids this.

For more information on this dependency, see
https://github.com/paulrobertlloyd/jekyll-figure.