Syntax highlight operators for select languages
Adds syntax highlighting for operator tokens on languages with the
exception of Java. Unfortunately, the Java lexer does not distinguish
between general punctuation (e.g. parentheses) and operators (e.g.
arithmetic operators), so highlighting operator tokens in Java would be
visually excessive. Currently, the list of languages for which operator
highlighting is enabled is simply all referenced languages in the site,
excluding Java.