Use system-ui generic font family in font stack
Replaces a couple platform-specific font families with the more general
alternative.

This also seems to fix what appears to be a bug in the Brave browser,
where the site rendered with "shields up" skips over the intended
BlinkMacSystemFont. I was unable to track down the root cause of this
issue but troubleshooting revealed that it appears to have to do with
the Brave "shield".