Add new custom Kramdown syntax highlighter
Adds and exports a new custom Kramdown syntax highlighter, named the
'custom_highlighter_delegator', which can be used to switch between
multiple syntax highlighters when highlighting code. Currently, only one
true syntax highlighter is supported (i.e. Rouge), but there is also a
fallback syntax highlighter called NoHighlight which the delegator can
use in the case where the requested highlighter is unknown.

As of this commit, this new highlighter is not yet enabled for the
project.