Add WIP Tree-sitter Ruby binding library
Adds a new Rust library, "Tree-sitter Ruby binding", to the project.
This library will eventually allow Ruby code to invoke the Rust binding
of the Tree-sitter highlight library.

As of this commit, this library is only able to HTML-escape code blocks,
and does not actually interface with Tree-sitter. This library is also
not currently used in building the site.

See https://tree-sitter.github.io/tree-sitter for more information about
Tree-sitter.