Use Tree-sitter loader library to load parsers
Replaces the current approach of loading language parsers using Rust
crates and statically defined highlight configurations with the
Tree-sitter loader library, which loads language parsers from the
file system at runtime.

The loader library has a number of advantages (and some slight
disadvantages, but I believe these are outweighed by its advantages)
over the previous approach. In particular, it outsources language
highlight configuration to Tree-sitter itself and, most importantly, it
supports more languages.

As part of these changes, the tree-sitter-loader crate was of course
added as a dependency, while the tree-sitter-python crate was removed.

For more information on tree-sitter-loader, see
https://crates.io/crates/tree-sitter-loader.