Add functionality for comparing highlighters
Adds new custom HTML converter functionality that allows for creating
bespoke HTML elements for comparing syntax highlighters against a given
code block. In particular, if 'highlighter="compare"' is set as a
Kramdown parameter for a given code block then, instead of rendering a
normal code block, a tab group will be rendered, where each tab shows
the effect of highlighting the code with a particular syntax
highlighter. The two syntax highlighters that are supported are Rouge
and Tree-sitter.

Also adds SCSS styling for the tab group so that it renders nicely.