Update name of hidden language descriptor class
Updates the name of the HTML class attribute for hiding the language
descriptor of a code block from 'language-descriptor-hidden' to
'lang-descriptor-hidden'.

This change fixes a bug where if this class is added to a code block
without an explicit language then Kramdown erroneously uses this class
to determine that the language of the code is 'descriptor-hidden'. This
is due to the fact that Kramdown looks for the string 'language-' in the
list of class names for a code block in order to determine the language,
which can of course cause problems if this string appears in the class
list for other reasons. The solution is to just use a different prefix
in the name of this particular class, namely 'lang-'.

Whether or not this behavior exhibited by Kramdown amounts to a bug or
just a sharp edge be weary of, I am not sure. I may eventually open an
issue with Kramdown to point out this behavior.