Use older version of Tree-sitter JavaScript parser
Reverts the Tree-sitter JavaScript submodule to use an older version of
the library. This revert is done so that this library can be used as a
dependency by the Tree-sitter TypeScript parser.

Also removes some token classes as these classes are not present in this
earlier version of the library.

Based on my testing, this older version (tagged 'rust-0.20.0') is
(probably the latest version) compatible with the latest version of the
Tree-sitter TypeScript parser library, whereas the version reverted from
is not.