Disable Dependabot updates for JavaScript parser
Disables Dependabot updates for the Tree-sitter JavaScript parser Git
submodule.

The Tree-sitter TypeScript parser submodule depends on the JavaScript
parser submodule and currently, the former is not compatible with newer
versions of the latter, hence the need for this change. With hope, these
circumstances will eventually change, at which point this change can be
reverted.