Use Kramdown Tree-sitter plugin for highlighting
Replaces the internal Tree-sitter plugin with the
kramdown-syntax_tree_sitter plugin for highlighting code with
Tree-sitter. As part of this change, some configuration settings were
added to the Jekyll configuration file, and the
CustomHighlighterDelegator highlighter was updated to adapt the output
of the new plugin.

More information on the plugin can be found here:
https://github.com/andrewtbiehl/kramdown-syntax_tree_sitter