Reorder Rust install steps in deployment workflow
It seems that the order of these steps, prior to these changes, doesn't
make much sense. Instead, it makes sense to me currently that first the
correct version of Rust should be installed, and then all cached Cargo
packages should be restored.

This change is also not particulaly relevant, as much of this will soon
be removed entirely.