Move Rust version configuration to deployment workflow
Moves the location defining the Rust version to use for building the
site from the tree_sitter_ruby_adapter project manifest to the GitHub
Actions deployment workflow.

This change was made because the project no longer depends on the
tree_sitter_ruby_adapter internal Rust project. The workflow
configuration file is not an ideal location for this information (in my
opinion), but it is perfectly fine for the time being.