Fix hypothetical bug in code snippet
Fixes a hypothetical bug in the Tree-sitter self-contained code
highlighting snippet by simplifying the behavior of the snippet.

The buggy line of code that was removed was intended to allow a user to
highlight embedded languages, however this is only possible if the
embedded language uses the exact same highlighting tokens as the
surrounding language, which is usually not true. A much more complicated
fix can be applied that enables highlighting of embedded languages but,
since this is supposed to be a minimal example, it made more sense to
simply remove this feature altogether. The resulting code still cannot
highlight embedded languages, but will no longer crash when attempting
to do so.