Disable Dependabot updates for Git submodules
This reverts commit ecd2c78456a7b5f0ea583a54c949307dcc3eaf55.

As of this commit, this project uses Git submodules solely for managing
Tree-sitter language parser dependencies. These Git repositories are
updated frequently, which leads to frequent Dependabot pull requests,
which are arguably not worth the effort to merge. Hence, automatic
Dependabot updates for these dependencies are being indefinitely
frozen. If these dependencies are worth keeping up-to-date at all, my
current opinion is that infrequent manual updates are preferable.